+++ src/domain/categories/dto/categories_dto.py
@@ -1,4 +1,4 @@
-from pydantic import BaseModel
+from ....common.validators import BaseModel
 
 
 class CategoryListResponse(BaseModel):
