+++ src/domain/scraping/scraping_module.py
@@ -2,7 +2,7 @@ from nest.core import Module
 from .scraping_service import ScrapingService
 from .scraping_controller import ScrapingController
 from .book_scraper import BookScraper
-from ...infra.repositories.book_repository_module import BookRepositoryModule
+from ...infra.repositories.book.book_repository_module import BookRepositoryModule
 
 
 @Module(
